requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
facenet-pytorch>=2.5.3
insightface>=0.7.3
onnxruntime>=1.16.0
scikit-learn>=1.3.0
scipy>=1.11.0
mediapipe>=0.10.0
pydantic>=2.5.0
aiosqlite>=0.19.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
facenet-pytorch>=2.5.3
insightface>=0.7.3
onnxruntime>=1.16.0
scikit-learn>=1.3.0
scipy>=1.11.0
mediapipe>=0.10.0
pydantic>=2.5.0
aiosqlite>=0.19.0
